refactor(analyzers): consolidate vision package into claude package
Move all vision analysis (screenshot, login verification, terminal reading)
from internal/analyzers/vision/ into internal/analyzers/claude/, unifying
the Claude API client into a single package.
