+++ internal/analyzers/claude/types.go
@@ -1,7 +1,7 @@
 // Copyright 2026 Praetorian Security, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package vision
+package claude
 
 import "context"
